Originating revision: r455 | cjplummer | 2006-11-21 14:48:15 -0800 (Tue, 21 Nov 2006) | 26 lines
Merged cyclops-oss-code-readines branch 445:446 into trunk:
                                                                                
Create the binary bundle as a zip file instead of a .tar.gz file.
                                                                                
Copy btclasses.zip into the lib directory of the binary bundle.
                                                                                
Support allowing other makefiles to append to the list of files or directories that end up in the binary bundle. This is done by adding to BINARY_BUNDLE_PATTERNS.
                                                                                
You can now specify a root/base directory that the binary bundle will unzip into.
                                                                                
The repository "legal" directory is now included in the binary bundle. By default an svn checkout of the files will be done. You also have the option of setting JAVAME_LEGAL_DIR to the location of an already checked out version rather than
having it do the svn checkout.
                                                                                
Support appending the svn revision number to the end of the binary bundle name.
                                                                                
These are the new options for building the binary bundle (which is done by using the "bin" target):
                                                                                
 - JAVAME_LEGAL_DIR: directory containing legal documents. If not set,
        the makefiles will automatically do an svn checkout of the files.
        from the respository. This will result in a password prompt if you
        haven't already authenticated with the repository.
 - BINARY_BUNDLE_NAME: name of the binary bundle, excluding .zip extension.
 - BINARY_BUNDLE_DIRNAME: directory name the binary bundle will unzip
        into. Defaults to $(BINARY_BUNDLE_NAME).
 - BINARY_BUNDLE_APPEND_REVISION(true): Appends the repository
        revision number to the end of BINARY_BUNDLE_NAME.