These two files should have been part of the previous commit with this message:
First implementation of the SimpleBasisAMS JUMP presentation mode.  A "presentation mode", a term which may or may not exist in the future, refers to an AMS type of application which uses JUMP APIs while running in the executive process.  A presentation mode usually contains an interactive UI, but a graphical UI is not required.

The provided SimpleBasisAMS can be used as a base presentation mode class which can be extended for more advanced features.  Note, the functionality of SimpleBasisAMS is still evolving and deliberately has no specification.  It is provided as a simple way to launch, kill, and switch between JUMP applications.  Currently, the only JUMP applications which can be run are ones installed with the JUMPInstallerTool, which is located in the installer module and has instructions for it's use within the JUMPInstallerTool.java file.

To create a presentation mode within JUMP:

1) Create a class which implements the com.sun.jump.presentation.JUMPLauncher interface.
2) Set the jump.launcher property within build/modules-config.properties to this class, including full package name.
3) Run the JUMP server and the JUMP executive, the presentation mode class will automatically run within the JUMP executive process.

To use the provided SimpleBasisAMS presentation mode:

1) Create a content store repository and use the JUMPInstallerTool (see instructions within JUMPInstallerTool.java ) to install applications.

2) Leave the jump.launcher property in build/modules-config.properties to com.sun.jumpimpl.presentation.simplebasis.SimpleBasisAMS

To not use any presentation mode by default when the JUMP executive starts:

1) Remove or comment out the jump.launcher property in build/modules-config.properties